requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
cachetools==5.3.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
exceptiongroup==1.1.3
fonttools==4.43.1
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
install==1.3.5
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
loguru==0.7.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
minify_html==0.11.1
multidict==6.0.4
networkx==3.1
numpy==1.25.1
openai==0.27.8
outcome==1.2.0
packaging==23.2
-e git+https://github.com/DatapaloozaCO/datapalooza_scraper_framework.git@6c63022e3d131f04861ca4d8b66f6f00ffb461d8#egg=palooza_wizard
pandas==2.0.3
Pillow==10.1.0
protobuf==4.24.4
pyarrow==13.0.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
selenium==4.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
streamlit==1.28.0
tenacity==8.2.3
tiktoken==0.5.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
trubrics==1.6.2
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.3
validators==0.22.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
